+++ loader/launcher/windows/fakeGeode.c
@@ -1,2 +1,1 @@
-#include <Windows.h>
-__declspec(dllexport) DWORD WINAPI loadGeode(void* arg) { return 0; }
+__declspec(dllexport) void fake() { }